refactor: seperate AddRecord component
